prov/verbs: remove odd verbs_av.c file
All content is moved to rdm/verbs_av_ep_rdm.c file
