+++ prov/verbs/Makefile.include
@@ -4,7 +4,6 @@ _verbs_files = \
 	prov/verbs/src/fi_verbs.c \
 	prov/verbs/src/uthash.h \
 	prov/verbs/src/verbs_atomic.c \
-	prov/verbs/src/verbs_av.c \
 	prov/verbs/src/verbs_cm.c \
 	prov/verbs/src/verbs_cq.c \
 	prov/verbs/src/verbs_srq.c \
